+++ packages/terra-form-select/src/_Menu.jsx
@@ -176,6 +176,11 @@ class Menu extends React.Component {
         intl,
       } = this.context;
 
+      // Race condition can occur between calling timeout and unmounting this component.
+      if (!visuallyHiddenComponent || !visuallyHiddenComponent.current) {
+        return;
+      }
+
       if (hasNoResults && searchValue) {
         visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.noResults' }, { text: searchValue });
       } else if (hasMaxSelection) {